+++ R/zzz.R
@@ -2,7 +2,7 @@
   ## set options using the approach used by devtools
   opts <- options()
   opts.landisutils <- list(
-    ## none yet
+    landisutils.cache.path = .climateCachePath()
   )
   toset <- !(names(opts.landisutils) %in% names(opts))
   if (any(toset)) options(opts.landisutils[toset])