+++ tests/testthat/test-ext_biomass_succession.R
@@ -68,8 +68,8 @@ testthat::test_that("Biomass Succession inputs are properly created", {
     .f = prep_daily_weather,
     studyArea = ecoregionPolys,
     id = "ecoregion",
-    start = glue::glue("{head(clim_years, 1)}-01-01"),
-    end = glue::glue("{tail(clim_years, 1)}-12-31")
+    start = head(clim_years, 1),
+    end = tail(clim_years, 1)
   ) |>
     purrr::list_rbind()
 
@@ -94,8 +94,8 @@ testthat::test_that("Biomass Succession inputs are properly created", {
     .f = prep_monthly_weather,
     studyArea = ecoregionPolys,
     id = "ecoregion",
-    start = glue::glue("{head(clim_years, 1)}-01-01"),
-    end = glue::glue("{tail(clim_years, 1)}-12-31")
+    start = head(clim_years, 1),
+    end = tail(clim_years, 1)
   ) |>
     purrr::list_rbind() |>
     dplyr::filter(Year <= tail(clim_years, 1)) ## match end year