+++ NEWS.md
@@ -1,3 +1,9 @@
+# landisutils (development version)
+
+* use cache with weather data acquisition (#1);
+* allow setting cache path using option `landisutils.cache.path`;
+* change arguments `start` and `end` in `prep_*_weather()` to be integer years;
+
 # landisutils 0.0.1
 
 * initial version;